services/vision/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
httpx==0.26.0
redis>=5.0,<6
python-multipart==0.0.9
psycopg2-binary==2.9.9
numpy>=1.24,<3
pandas>=2.0,<3
scipy>=1.10,<2
statsmodels>=0.14,<1
scikit-learn>=1.3,<2
pytest==8.0.2
pytest-asyncio==0.23.5
